Fix one occurrence of corrupt trees.
Caused by usage of stale parent.tree_id when inserting nodes as direct children of root nodes.

This fixes one of the causes of #320 but there are others, e.g. this same thing still happens for non-root nodes.

PERFORMANCE: This necessitates an extra query to get the appropriate tree_id value when inserting.
If this is a problem when inserting lots of nodes, consider using `delay_mptt_updates` for bulk insertion.