Don't import south unless it's in installed apps
South is often included in 'install_requires' for 3rd party apps even for Django 1.7+

See https://github.com/spookylukey/django-paypal/issues/101 for an explanation as to why. 

South obviously won't be updated to support Django 1.7+ so this will cause warnings and errors if it's imported. Therefore it's much safer to check for South being in INSTALLED_APPS